# src/main/java/com/rabbit/api/dto/request/RegisterRequest.java
package com.rabbit.api.dto.request;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class RegisterRequest {

    @NotNull(message = "User details are required")
    @Valid
    private CreateUserRequest user;
}