# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Invoice API
  description: |
    Multi-tenant SaaS API for invoice management.

    ## Authentication
    All protected endpoints require `Authorization: Bearer <token>` header.

    Multi-company access uses `X-Company-Id` header to select the active company context.

    ## Roles
    - **SUPER_ADMIN** — Full access. Can create companies, add users, manage everything.
    - **ADMIN** — Can view and manage document types.
    - **USER** — Basic access within a company.

    ## Getting Started
    1. Register via `POST /auth/register`
    2. Verify email via `POST /auth/verify`
    3. Login via `POST /auth/login` to get JWT token
    4. Use token in `Authorization: Bearer <token>` header
    5. Create a company via `POST /companies`
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - Bearer Authentication: []
components:
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token obtained from login
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/CreateUserRequest'
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
        - city
        - state
        - pinCode
        - country
      properties:
        username:
          type: string
          description: Unique username
          example: johndoe
        email:
          type: string
          format: email
          description: Valid email address
          example: john@example.com
        password:
          type: string
          description: "Min 8 chars with uppercase, lowercase, digit, and special character"
          example: Password@1
        firstName:
          type: string
          description: First name
          example: John
        lastName:
          type: string
          description: Last name
          example: Doe
        addressLine1:
          type: string
          description: Address line 1
          example: 123 Main Street
        addressLine2:
          type: string
          description: Address line 2
          example: Suite 4B
        city:
          type: string
          description: City
          example: Mumbai
        state:
          type: string
          description: State
          example: Maharashtra
        pinCode:
          type: string
          description: PIN code (6 digits)
          pattern: "^[0-9]{6}$"
          example: "400001"
        country:
          type: string
          description: Country
          example: India
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Verification token from email
          example: 550e8400-e29b-41d4-a716-446655440000
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email address or username
          example: john@example.com
        password:
          type: string
          description: Account password
          example: Password@1
...
    RegisterResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Username
          example: johndoe
        email:
          type: string
          format: email
          description: Email address
          example: john@example.com
        firstName:
          type: string
          description: First name
          example: John
        lastName:
          type: string
          description: Last name
          example: Doe
        status:
          type: string
          description: Account status
          enum:
            - INACTIVE
            - PENDING_VERIFICATION
            - ACTIVE
            - LOCKED
            - DISABLED
          example: INACTIVE
        message:
          type: string
          description: Status message
          example: Verification email sent. Please verify your email.
    VerifyEmailResponse:
      type: object
      properties:
        status:
          type: string
          description: Verification status
          example: VERIFIED
        message:
          type: string
          description: Status message
          example: Email verified successfully
    LoginResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Username
          example: johndoe
        email:
          type: string
          format: email
          description: Email address
          example: john@example.com
        firstName:
          type: string
          description: First name
          example: John
        lastName:
          type: string
          description: Last name
          example: Doe
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiJ9...
        expiresIn:
          type: integer
          format: int64
          description: Token expiry in milliseconds
          example: 86400000
    UserMeResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Username
          example: johndoe
        email:
          type: string
          format: email
          description: Email address
          example: john@example.com
        firstName:
          type: string
          description: First name
          example: John
        lastName:
          type: string
          description: Last name
          example: Doe
        city:
          type: string
          description: City
          example: Mumbai
        state:
          type: string
          description: State
          example: Maharashtra
        country:
          type: string
          description: Country
          example: India
        status:
          type: string
          description: Account status
          enum:
            - INACTIVE
            - PENDING_VERIFICATION
            - ACTIVE
            - LOCKED
            - DISABLED
          example: ACTIVE
        companies:
          type: array
          description: Companies the user has access to
          items:
            $ref: '#/components/schemas/CompanyInfo'
    CompanyInfo:
      type: object
      description: Company summary in user profile
      properties:
        companyId:
          type: string
          format: uuid
          description: Company UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Company name
          example: Acme Corp
        role:
          type: string
          description: User's role in this company
          enum:
            - SUPER_ADMIN
            - ADMIN
            - USER
          example: SUPER_ADMIN
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2026-01-01T00:00:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Validation failed
        path:
          type: string
          description: Request path
          example: /api/v1/auth/register
        validationErrors:
          type: object
          description: Field-level validation errors (only for 400 validation errors)
          additionalProperties:
            type: string
          example:
            email: Email should be valid
            password: Password must be at least 8 characters
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Custom form registration. Creates a new user account in PENDING_VERIFICATION status
        and sends a verification email. User must verify email before logging in.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: |
        Verifies user email using the token sent to their email.
        Account becomes ACTIVE after verification.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmailResponse'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email or username
      description: |
        Authenticates user with email or username and password.
        Returns JWT token. User must be verified to login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Returns the authenticated user's profile and company list.
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        '401':
          description: Token not provided or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend verification email
      description: Sends a new verification email to the user. Only works for unverified accounts.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: Registered email address
                  example: john@example.com
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Verification email sent
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Sends a password reset email to the user. Works for both verified and unverified accounts. Unverified accounts will also be verified upon password reset.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: Registered email address
                  example: john@example.com
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset email sent
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password using token
      description: |
        Resets the user's password using the token from the reset email.
        If the user was not verified, they will also be verified upon successful password reset.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                  description: Password reset token from email
                  example: 550e8400-e29b-41d4-a716-446655440000
                newPassword:
                  type: string
                  description: "Min 8 chars with uppercase, lowercase, digit, and special character"
                  example: NewPassword@1
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset successfully
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...